When rendering to postscript, convert text to paths
For whatever reason, the postscript 3up-ing process messed up the
rendering of the name field on the first middle-of-page ticket.  It
also only seemed to occur for certain character sequences (I tested
with "def").  Instead of faffing around trying to work out what was
going wrong in ghostscript somewhere, I've just converted the text
to paths instead.  This is not a problem really -- tickets will be
printed.