Frob the user ticket scanner.
This is an ugly patch sadly. It eliminates all checking that users are
in the media consent group from the ticket scanner: this is un-necessary
because they should only have been issued a ticket if the user is in
that group. Retrieving that information from badger for the user scanner
is over-checking.

It also removes operations in scanner.py that use the 'User' object:
again, this data isn't accessible on the scanning machine under the new
model.

The look-up-user-from-badger feature is deliberately broken: as
discussed on srobo-devel, this will go via nemesis in the future (and
will be as small as possible).

Finally, the user-data-fetching stuff in user_database.py become static
methods, as they don't interact with user objects any more.