Access the ticket system via localhost
The ticket system is no longer hosted on the same machine as handles
the root of the domain, so the previous redirect issues which were
seen here are no longer an issue. This avoids at least an external
network round-trip as well as making the ticket system work on
development boxes too.